+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-### Unreleased
+### [v0.15](https://github.com/sting-ioc/sting/tree/v0.15) (2020-04-23) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.14...v0.15) · [API Differences](https://sting-ioc.github.io/api-diff?key=sting&old=0.14&new=0.15)
+
+The release includes 1 non breaking API change.
+
+Changes in this release:
 
 * Upgrade the `org.realityforge.proton` artifacts to version `0.46`.
 * Generate an error if a `@Fragment` annotated type includes itself.